refactor styles into new files
